refactor: move ingest helpers into dataregistry so the router doesn't import scripts/
dataregistry/api/kpn_cms.py imported scripts.import_kpn_cms.rows_for, but the
Docker image only ships dataregistry/ (Dockerfile only COPYs that directory),
so the documented docker run path 500ed with ModuleNotFoundError on every
proxy-persist. It was also an architectural inversion: the API depending on
an ops CLI under scripts/.

Move rows_for, strip_html, and _field_value (plus the _TAG_RE regex they use)
into a new dataregistry/api/kpn_cms_ingest.py. scripts/import_kpn_cms.py now
imports them from there and re-exports rows_for/strip_html so its existing
public interface (and the tests that import from it) is unchanged;
_harvest_nids and its _NODE_RE regex stay in the script since they're only
used there. kpn_cms.py now imports rows_for from dataregistry.api.kpn_cms_ingest
directly, with no scripts/ dependency left in the API.

Also add `COPY scripts ./scripts` to the Dockerfile (the import CLI is part
of the ops story and should ship in the image), and drop the ASSET_PREFIX
import from scripts/import_kpn_cms.py, which the move left genuinely unused.
